[refactor] 불필요한 변수 삭제
